--- Quiz-Game/index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>Quiz-Game</title>
-
-</head>
-<body>
-    <h1>Quiz-game</h1>
-    <nav id="nav">
-        <p id="scoreHolder"> </p>
-    </nav>
-    <main class="container">
-        <section class="sub-container">
-            <div id="holder">
-                
-            </div>
-        </section>
-    </main>
-
-    <audio id="sound3" src="./sound/sounds_boom (1).wav"></audio>   
-    <audio id="sound2" src="./sound/sounds_kick.wav"></audio>
-    <audio id="sound1" src="./sound/sounds_tink (1).wav"></audio>
-
-    <script type="module" src="./script/question.js"></script>
-    <script type="module" src="./script/score.js"></script>
-</body>
-</html>
