+++ Quiz-Game/quiz.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>Quiz-Game</title>
+
+</head>
+<body>
+   
+   <main id="container">
+    <section id="sub-container1">
+        <div id="choices">
+            <button id="game">Game</button>
+            <button id="music">music</button>
+            <button id="film">film</button>
+        </div>
+        
+        <div id="holder"></div>
+    </section>
+    <section id="sub-container2">
+            <aside id="nav">
+                <p id="scoreHolder"> </p>
+            </aside>
+            <aside>
+                <div id="resultHolder"></div>
+            </aside>
+     </section>
+  </main>
+
+    <!--Select sound-->
+    <audio id="sound3" src="./sound/sounds_boom (1).wav"></audio>   
+    <audio id="sound2" src="./sound/sounds_kick.wav"></audio>
+    <audio id="sound1" src="./sound/sounds_tink (1).wav"></audio>
+    <!--link script file-->
+    <script type="module" src="./src/question.js"></script>
+    <script type="module" src="./src/score.js"></script>
+</body>
+</html>
