--- Quiz-Game/script/fail.js
@@ -1,5 +1,0 @@
-export function fail(){
-    const sound3 = document.getElementById("sound3")
-    console.log("fail")
-    sound3.play()
-  }
