--- Quiz-Game/script/success.js
@@ -1,6 +1,0 @@
-export function success(){
-   const sound1 = document.getElementById("sound1")
-    console.log("success")
-    sound1.play()
-
-}
