+++ Quiz-Game/src/feedback.js
@@ -1,0 +1,11 @@
+export function success(){
+  
+   const sound1 = document.getElementById("sound1")
+    console.log("success")
+    sound1.play()
+}
+export function fail(){
+    const sound3 = document.getElementById("sound3")
+    console.log("fail")
+    sound3.play()
+  }
