+++ Quiz-Game/src/score.js
@@ -2,6 +2,7 @@ import { score } from "./question.js";
 
 export function getSum(){
     const scoreHolder = document.getElementById("scoreHolder")
+    scoreHolder.id = "scoreHolder"
     const sum = score.reduce((a, b) => a + b, 0);
     scoreHolder.innerHTML = sum
 }
