+++ Quiz-Game/style.css
@@ -1,6 +1,47 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+#container{
+    display: grid;
+    grid-template-columns: 80% 20%;
+}
+#sub-container1{
+    margin: 5% 5% 5% 20%;
+}
+#sub-container2{
+    margin: 0% 0% 5% 20%;
+    height: auto;
+    min-height: 690px;
+    background-color: rgb(105, 99, 99);
+}
+#answerInput{
+    width: 50%;
+    height: 40px;
+}
+#submitButton{
+    width: 10%;
+    height: 50px;
+    margin: 2%;
+}
+#successDiv{
+    padding: 5%;
+    margin: 3%;
+}
+#failDiv{
+    padding: 5%;
+    margin: 3%;
+}
+#scoreHolder{
+    text-align: center;
+    padding: 5%;
+}
 #successDiv{
     background-color: green;
 }
 #failDiv{
     background-color: red;
+}
+#questionLine{
+    font-size: 20px;
 }
